Cimtest: Check for None return value
Some tests were raising Attribute Exception due to None value returned.

Changes from #1:
    - Explicitly check for None if value returned from GetInstance()

+++ suites/libvirt-cim/cimtest/VirtualSystemManagementService/14_define_sys_disk.py
@@ -154,7 +154,7 @@ def main():
         logger.error(details)
         status = FAIL
 
-    if os.path.exists(addr):
+    if addr and os.path.exists(addr):
         os.remove(addr)
 
     if guest_defined == True:
